+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Unable to enable or disable modules
+- Module list failing to output after a module has been disable and re-enabled
+
 ## [0.3.0] - 2020-02-08
 ### Added
 - `queue` subcommands added to modify the current songs in the queue